completed

+++ Services/StudentService.py
@@ -18,7 +18,17 @@ while True:
     menu()
     choice = input("Vui lòng chọn một chức năng: ")
     if choice == "1":
-        studentDAO.findAll()
+        print("==== List Student ====")
+        students = studentDAO.findAll()
+        s = '| %s | %s | %s %s | %s | %s | %s | %s |'
+        head = ['ID', 'Mã sinh viên', 'Họ', 'tên', 'Ngày sinh', 'Điểm toán', 'Điểm lý', 'Điểm hóa']
+        print('| {:3s} | {:10s} | {:20s} | {:10s} | {:10s} | {:10s} | {:10s} | {:10s} |'
+              .format('ID', 'Mã sinh viên', 'Họ', 'tên', 'Ngày sinh', 'Điểm toán', 'Điểm lý', 'Điểm hóa'))
+        # print (s % tuple(head))
+        for student in students:
+            # print (s % tuple(student))
+            print('| {:3d} | {:12s} | {:20s} | {:10s} | {:10s} | {:10.2f} | {:10.2f} | {:10.2f} |'
+                  .format(student[0], student[1], student[2], student[3], student[4].strftime("%d-%m-%Y"), student[5], student[6], student[7]))
     elif choice == "2":
         code = input("Nhập mã sinh viên: ")
         firstName = input("Nhập họ và tên đệm: ")
@@ -34,7 +44,9 @@ while True:
         if student is None:
             print("Không tìm thấy sinh viên")
         else:
-            print(student)
+            print('| {:3d} | {:12s} | {:20s} | {:10s} | {:10s} | {:10.2f} | {:10.2f} | {:10.2f} |'
+                  .format(student[0], student[1], student[2], student[3], student[4].strftime("%d-%m-%Y"), student[5],
+                          student[6], student[7]))
             code = input("Nhập mã sinh viên: ")
             firstName = input("Nhập họ và tên đệm: ")
             lastName = input("Nhập tên: ")
@@ -44,17 +56,34 @@ while True:
             chemistry = float(input("Nhập điểm hóa: "))
 
             studentDAO.update((code, firstName, lastName, dob, math, physics, chemistry, studentId))
-
+            print("After updating record ")
+            studentAfterUpdated = studentDAO.findById(studentId)
+            print(studentAfterUpdated)
     elif choice == "4":
         studentId = input('Nhập id muốn xoá: ')
-        studentDAO.delete(studentId)
+        student = studentDAO.findById(studentId)
+        if student is None:
+            print("Không tìm thấy sinh viên")
+        else:
+            print('| {:3d} | {:12s} | {:20s} | {:10s} | {:10s} | {:10.2f} | {:10.2f} | {:10.2f} |'
+                  .format(student[0], student[1], student[2], student[3], student[4].strftime("%d-%m-%Y"), student[5],
+                          student[6], student[7]))
+
+            print('Bạn có chắc chắn xóa thông tin này ?')
+            print('1.Vâng')
+            print('2.Hủy')
+            confirm = int(input("Vui lòng chọn 1 trong 2: "))
+            if confirm == 1:
+                studentDAO.delete(studentId)
     elif choice == "5":
         studentId = input('Nhập id muốn tìm: ')
         student = studentDAO.findById(studentId)
         if student is None:
             print("Không tìm thấy sinh viên")
         else:
-            print(student)
+            print('| {:3d} | {:12s} | {:20s} | {:10s} | {:10s} | {:10.2f} | {:10.2f} | {:10.2f} |'
+                  .format(student[0], student[1], student[2], student[3], student[4].strftime("%d-%m-%Y"), student[5],
+                          student[6], student[7]))
     elif choice == "6":
         exit_program()
     else:
